Cleanup utils to prepare for move.
