+++ ocean_provider/serializers.py
@@ -4,12 +4,8 @@
 #
 import json
 
-from ocean_provider.util import get_asset_url_at_index
-from ocean_provider.utils.basics import get_asset_from_metadatastore, get_config
-
-
-def get_metadata_url():
-    return get_config().aquarius_url
+from ocean_provider.util import get_asset_url_at_index, get_metadata_url
+from ocean_provider.utils.basics import get_asset_from_metadatastore
 
 
 class StageAlgoSerializer:
