+++ ocean_provider/utils/test/test_utils.py
@@ -3,8 +3,17 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 import logging
+import mimetypes
+from copy import deepcopy
+from unittest.mock import MagicMock, Mock
 
-from ocean_provider.util import service_unavailable
+from ocean_lib.common.http_requests.requests_session import get_requests_session
+from ocean_provider.util import (
+    build_download_response,
+    get_download_url,
+    service_unavailable,
+)
+from werkzeug.utils import get_content_type
 
 test_logger = logging.getLogger(__name__)
 
@@ -18,3 +27,112 @@ def test_service_unavailable(caplog):
     assert response["error"] == "test message"
     assert response["context"] == context
     assert caplog.records[0].msg == "Payload was: item1=test1,item2=test2"
+
+
+def test_build_download_response():
+    request = Mock()
+    request.range = None
+
+    class Dummy:
+        pass
+
+    mocked_response = Dummy()
+    mocked_response.content = b"asdsadf"
+    mocked_response.status_code = 200
+    mocked_response.headers = {}
+
+    requests_session = Dummy()
+    requests_session.get = MagicMock(return_value=mocked_response)
+
+    filename = "<<filename>>.xml"
+    content_type = mimetypes.guess_type(filename)[0]
+    url = f"https://source-lllllll.cccc/{filename}"
+    response = build_download_response(request, requests_session, url, url, None)
+    assert response.headers["content-type"] == content_type
+    assert (
+        response.headers.get_all("Content-Disposition")[0]
+        == f"attachment;filename={filename}"
+    )
+
+    filename = "<<filename>>"
+    url = f"https://source-lllllll.cccc/{filename}"
+    response = build_download_response(request, requests_session, url, url, None)
+    assert response.headers["content-type"] == get_content_type(
+        response.default_mimetype, response.charset
+    )
+    assert (
+        response.headers.get_all("Content-Disposition")[0]
+        == f"attachment;filename={filename}"
+    )
+
+    filename = "<<filename>>"
+    url = f"https://source-lllllll.cccc/{filename}"
+    response = build_download_response(
+        request, requests_session, url, url, content_type
+    )
+    assert response.headers["content-type"] == content_type
+
+    matched_cd = (
+        f"attachment;filename={filename+mimetypes.guess_extension(content_type)}"
+    )
+    assert response.headers.get_all("Content-Disposition")[0] == matched_cd
+
+    mocked_response_with_attachment = deepcopy(mocked_response)
+    attachment_file_name = "test.xml"
+    mocked_response_with_attachment.headers = {
+        "content-disposition": f"attachment;filename={attachment_file_name}"
+    }
+
+    requests_session_with_attachment = Dummy()
+    requests_session_with_attachment.get = MagicMock(
+        return_value=mocked_response_with_attachment
+    )
+
+    url = "https://source-lllllll.cccc/not-a-filename"
+    response = build_download_response(
+        request, requests_session_with_attachment, url, url, None
+    )
+    assert (
+        response.headers["content-type"]
+        == mimetypes.guess_type(attachment_file_name)[0]
+    )  # noqa
+
+    matched_cd = f"attachment;filename={attachment_file_name}"
+    assert response.headers.get_all("Content-Disposition")[0] == matched_cd
+
+    mocked_response_with_content_type = deepcopy(mocked_response)
+    response_content_type = "text/csv"
+    mocked_response_with_content_type.headers = {"content-type": response_content_type}
+
+    requests_session_with_content_type = Dummy()
+    requests_session_with_content_type.get = MagicMock(
+        return_value=mocked_response_with_content_type
+    )
+
+    filename = "filename.txt"
+    url = f"https://source-lllllll.cccc/{filename}"
+    response = build_download_response(
+        request, requests_session_with_content_type, url, url, None
+    )
+    assert response.headers["content-type"] == response_content_type
+    assert (
+        response.headers.get_all("Content-Disposition")[0]
+        == f"attachment;filename={filename}"
+    )
+
+
+def test_download_ipfs_file(client):
+    cid = "QmQfpdcMWnLTXKKW9GPV7NgtEugghgD6HgzSF6gSrp2mL9"
+    url = f"ipfs://{cid}"
+    download_url = get_download_url(url, None)
+    requests_session = get_requests_session()
+
+    request = Mock()
+    request.range = None
+
+    print(f"got ipfs download url: {download_url}")
+    assert download_url and download_url.endswith(f"ipfs/{cid}")
+    response = build_download_response(
+        request, requests_session, download_url, download_url, None
+    )
+    assert response.data, f"got no data {response.data}"
