+++ tests/test_download.py
@@ -3,30 +3,21 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 
-import mimetypes
-from copy import deepcopy
-from unittest.mock import MagicMock, Mock
-
 from ocean_lib.common.agreements.service_agreement import ServiceAgreement
 from ocean_lib.common.agreements.service_types import ServiceTypes
-from ocean_lib.common.http_requests.requests_session import get_requests_session
 from ocean_lib.models.data_token import DataToken
 from ocean_lib.web3_internal.transactions import sign_hash
 from ocean_lib.web3_internal.utils import add_ethereum_prefix_and_hash_msg
 from ocean_provider.constants import BaseURLs
-from ocean_provider.util import build_download_response, get_download_url
 from ocean_provider.utils.accounts import generate_auth_token
 from tests.test_helpers import (
-    get_consumer_wallet,
     get_dataset_ddo_with_access_service,
     get_dataset_with_invalid_url_ddo,
     get_dataset_with_ipfs_url_ddo,
     get_nonce,
-    get_publisher_wallet,
     mint_tokens_and_wait,
     send_order,
 )
-from werkzeug.utils import get_content_type
 
 
 def dummy_callback(*_):
@@ -103,115 +94,6 @@ def test_empty_payload(client):
 
 def test_exec_endpoint():
     pass
-
-
-def test_download_ipfs_file(client):
-    cid = "QmQfpdcMWnLTXKKW9GPV7NgtEugghgD6HgzSF6gSrp2mL9"
-    url = f"ipfs://{cid}"
-    download_url = get_download_url(url, None)
-    requests_session = get_requests_session()
-
-    request = Mock()
-    request.range = None
-
-    print(f"got ipfs download url: {download_url}")
-    assert download_url and download_url.endswith(f"ipfs/{cid}")
-    response = build_download_response(
-        request, requests_session, download_url, download_url, None
-    )
-    assert response.data, f"got no data {response.data}"
-
-
-def test_build_download_response():
-    request = Mock()
-    request.range = None
-
-    class Dummy:
-        pass
-
-    mocked_response = Dummy()
-    mocked_response.content = b"asdsadf"
-    mocked_response.status_code = 200
-    mocked_response.headers = {}
-
-    requests_session = Dummy()
-    requests_session.get = MagicMock(return_value=mocked_response)
-
-    filename = "<<filename>>.xml"
-    content_type = mimetypes.guess_type(filename)[0]
-    url = f"https://source-lllllll.cccc/{filename}"
-    response = build_download_response(request, requests_session, url, url, None)
-    assert response.headers["content-type"] == content_type
-    assert (
-        response.headers.get_all("Content-Disposition")[0]
-        == f"attachment;filename={filename}"
-    )
-
-    filename = "<<filename>>"
-    url = f"https://source-lllllll.cccc/{filename}"
-    response = build_download_response(request, requests_session, url, url, None)
-    assert response.headers["content-type"] == get_content_type(
-        response.default_mimetype, response.charset
-    )
-    assert (
-        response.headers.get_all("Content-Disposition")[0]
-        == f"attachment;filename={filename}"
-    )
-
-    filename = "<<filename>>"
-    url = f"https://source-lllllll.cccc/{filename}"
-    response = build_download_response(
-        request, requests_session, url, url, content_type
-    )
-    assert response.headers["content-type"] == content_type
-
-    matched_cd = (
-        f"attachment;filename={filename+mimetypes.guess_extension(content_type)}"
-    )
-    assert response.headers.get_all("Content-Disposition")[0] == matched_cd
-
-    mocked_response_with_attachment = deepcopy(mocked_response)
-    attachment_file_name = "test.xml"
-    mocked_response_with_attachment.headers = {
-        "content-disposition": f"attachment;filename={attachment_file_name}"
-    }
-
-    requests_session_with_attachment = Dummy()
-    requests_session_with_attachment.get = MagicMock(
-        return_value=mocked_response_with_attachment
-    )
-
-    url = "https://source-lllllll.cccc/not-a-filename"
-    response = build_download_response(
-        request, requests_session_with_attachment, url, url, None
-    )
-    assert (
-        response.headers["content-type"]
-        == mimetypes.guess_type(attachment_file_name)[0]
-    )  # noqa
-
-    matched_cd = f"attachment;filename={attachment_file_name}"
-    assert response.headers.get_all("Content-Disposition")[0] == matched_cd
-
-    mocked_response_with_content_type = deepcopy(mocked_response)
-    response_content_type = "text/csv"
-    mocked_response_with_content_type.headers = {"content-type": response_content_type}
-
-    requests_session_with_content_type = Dummy()
-    requests_session_with_content_type.get = MagicMock(
-        return_value=mocked_response_with_content_type
-    )
-
-    filename = "filename.txt"
-    url = f"https://source-lllllll.cccc/{filename}"
-    response = build_download_response(
-        request, requests_session_with_content_type, url, url, None
-    )
-    assert response.headers["content-type"] == response_content_type
-    assert (
-        response.headers.get_all("Content-Disposition")[0]
-        == f"attachment;filename={filename}"
-    )
 
 
 def test_asset_info(client):
